day lai

+++ Base365/controllers/qlts/taiSan.js
@@ -111,23 +111,23 @@ exports.showAll = async (req, res) => {
           ts_id: { $first: '$ts_id' },
           ts_ten: { $first: '$ts_ten' },
           nguoi_cam: { $first: '$id_ten_quanly' },
-          tong_so_luong: { $first: '$sl_bandau' },
-          so_luong_cap_phat: { $sum: { $cond: [{ $ifNull: ['$cap_phat.cp_da_xoa', 0] }, 0, 1] } },
-          so_luong_thu_hoi: { $sum: { $cond: [{ $ifNull: ['$thuhoi.xoa_thuhoi', 0] }, 0, 1] } },
+          tong_so_luong: { $addToSet: '$sl_bandau' },
+          so_luong_cap_phat: {
+            $sum: {
+              $cond: [{ $eq: ['$cap_phat.cp_da_xoa', 0] }, { $size: '$cap_phat.cap_phat_taisan.ds_ts.ts_id' }, 0],
+            },
+          },
+          so_luong_thu_hoi: {
+            $sum: {
+              $cond: [{ $eq: ['$thuhoi.xoa_thuhoi', 0] }, { $size: '$thuhoi.thuhoi_taisan.ds_thuhoi.ts_id' }, 0],
+            },
+          },
           so_luong_con_lai: { $first: '$soluong_cp_bb' },
           loai_ts: { $first: '$name_loai.ten_loai' },
           gia_tri: { $first: '$ts_gia_tri' },
           tinh_trang_su_dung: { $first: '$ts_trangthai' },
           don_vi_quan_ly: { $first: '$ts_don_vi' },
           vi_tri_tai_san: { $first: '$name_vitri.vi_tri' },
-        },
-      },
-      {
-        $match: {
-          $or: [
-            { so_luong_cap_phat: { $gt: 0 } },
-            { so_luong_thu_hoi: { $gt: 0 } },
-          ],
         },
       },
       {
@@ -235,7 +235,15 @@ exports.addTaiSan = async (req, res) => {
       return functions.setError(res, 'không có quyền truy cập', 400);
     }
     const validationResult = quanlytaisanService.validateTaiSanInput(ts_ten, ts_don_vi, id_dv_quanly, id_ten_quanly, id_loai_ts, ts_vi_tri);
-    const checkidNhom = await LoaiTaiSan.findOne({ id_loai: id_loai_ts, loai_da_xoa: 0 }).select('id_nhom_ts')
+    const checkidNhom = await LoaiTaiSan.findOne({ id_loai: id_loai_ts, loai_da_xoa: 0 ,id_cty : com_id}).select('id_nhom_ts')
+    if (!checkidNhom) {
+      // Xử lý lỗi hoặc thông báo không tìm thấy nhóm tài sản tương ứng
+      return functions.setError(res, 'Không tìm thấy nhóm tài sản tương ứng với id_loai_ts để thêm mới', 400);
+    }
+    let checkidVitri = await ViTri_ts.findOne({id_vitri : ts_vi_tri ,id_cty : com_id}).select('id_vitri')
+    if(!checkidVitri){
+      return functions.setError(res, 'Không tìm thấy vị trí tài sản tương ứng để thêm mới ', 400);
+    }
     if (validationResult === true) {
       
       let maxID = await quanlytaisanService.getMaxID(TaiSan);
@@ -255,7 +263,7 @@ exports.addTaiSan = async (req, res) => {
         soluong_cp_bb: soluong_cp_bb,
         ts_gia_tri: ts_gia_tri,
         ts_don_vi: ts_don_vi,
-        ts_vi_tri: ts_vi_tri,
+        ts_vi_tri: checkidVitri.id_vitri,
         ts_trangthai: ts_trangthai,
         ts_da_xoa: ts_da_xoa,
         ts_date_create: createDate,
@@ -333,32 +341,44 @@ exports.deleteTs = async (req, res) => {
       return functions.setError(res, 'không có quyền truy cập', 400);
     }
     const deleteDate = Math.floor(Date.now() / 1000);
+    if (!ts_id.every(num => !isNaN(parseInt(num)))) {
+      return functions.setError(res, 'ts_id không hợp lệ', 400);
+    }
     if (type == 1) { // xóa vĩnh viễn
       let idArraya = ts_id.map(idItem => parseInt(idItem));
-      await TaiSan.deleteMany({ ts_id: { $in: idArraya }, id_cty: com_id });
+      let result = await TaiSan.deleteMany({ ts_id: { $in: idArraya }, id_cty: com_id });
+      if (result.deletedCount === 0) {
+        return functions.setError(res, 'Không tìm thấy bản ghi phù hợp để xóa', 400);
+      }
       return functions.success(res, 'xóa thành công!');
     } else if (type == 0) {
       // thay đổi trạng thái là 1
       let idArray = ts_id.map(idItem => parseInt(idItem));
-      await TaiSan.updateMany(
-        { ts_id: { $in: idArray },ts_da_xoa: 0 },
+      let result = await TaiSan.updateMany(
+        { ts_id: { $in: idArray },ts_da_xoa: 0,id_cty : com_id },
         { ts_da_xoa: 1,
          ts_id_ng_xoa : ts_id_ng_xoa,
          ts_date_delete : deleteDate,
 
         }
       );
+      if (result.nModified === 0) {
+        return functions.setError(res, 'Không tìm thấy bản ghi phù hợp để thay đổi', 400);
+      }
       return functions.success(res, 'Bạn đã xóa thành công , hiện vào danh sách dã xóa !');
     } else if (type == 2) {
       // Khôi phục tài sản
       let idArray = ts_id.map(idItem => parseInt(idItem));
-      await TaiSan.updateMany(
+      let result = await TaiSan.updateMany(
         { ts_id: { $in: idArray }, 
-        ts_da_xoa: 1 },
+        ts_da_xoa: 1,id_cty : com_id  },
         { ts_da_xoa: 0 ,
           ts_id_ng_xoa : 0,
           ts_date_delete : 0,}
       );
+      if (result.nModified === 0) {
+        return functions.setError(res, 'Không tìm thấy bản ghi phù hợp để thay đổi', 400);
+      }
       return functions.success(res, 'Bạn đã khôi phục tài sản thành công!');
     } else {
       return functions.setError(res, 'không thể thực thi!', 400);
